// client/src/model/IIndex.ts
export interface Movie {
  id: string;
  title: string;
  year: string;
  genres: string[];
  ratings: number[];
  poster: string;
  contentRating: string;
  duration: string;
  releaseDate: string;
  averageRating: number;
  originalTitle: string;
  storyline: string;
  actors: string[];
  imdbRating: string;
  posterurl: string;
}